Clean up jflapgrader.py
The original version of this file was, quite frankly, atrocious. It had
tons of commented-out debugging statements; inconsistent indentation;
egregious spacing; some of the silliest comments I've ever seen in my
life; imports in the middle of the file (and nested inside functions!);
absolutely no consistent style conventions; a large commented-out
subscript at the bottom; and quite a few more problems to boot.

I've done my best to remove all these problems. The complete lack of
useful documentation, the obscenely chosen variable names, the
everything-as-globals programming style, and the fact that several
hundred lines are copy/pasted but changed subtly still make the code
rather formidable, but I tried to stick to non-semantic changes in this
commit.