Expand allowable syntax in test files
In particular, the original syntax (where only 'reject' was processed
and 'accept' was ignored, and the parsing was primarily driven by how
many spaces there were in the line(!)) has been thrown out completely in
favor of something more sane, where the following types of test cases
are allowable:

1010 -> accept
01 rej
empty->yes

Also, you can embed Python code directly into the test files, provided
that it takes the form of a function declaration for 'check' or 'words'.
The former function, if declared, will be used to compute the desired
result (True for accept and False for reject) for any test cases where
you don't specify the desired result, i.e.

empty
101

The latter function, if declared, should return a list of additional
inputs whose desired results are computed using 'check'. In addition to
the Python standard library, it has access to the 'all_bitstrings'
function, which returns a list of all possible bitstrings of length no
greater than its argument (in lexicographic order).

The exact format of test files will be documented in README.md in the
future.