Fix bug in tm_stateTrans2
It looked like the final 'return False' was unneeded, so I had removed
it in one of my previous cleanups. Unfortunately, the control flow is so
weird in this code that I missed a path, allowing tm_stateTrans2 to
return None in some cases where it was supposed to return False. This
should fix that bug.